Lottery Number Generator
This code will allow you to choose the lowest and the highest number, after that it will ask you how many numbers you want to draw. in order to your answers to its question it will generate a sequence of casual numbers that will appear on your display at the en of the process.